+++ examples/008-spi/f427vi/src/main.c
@@ -4,7 +4,7 @@
 // =============================================================================
 void taskPrint(void) {
     static int32_t loop = 220;
-    float data = -sin(( double )loop / 180.0 * 3.1415926);
+    double data = -sin(( double )loop / 180.0 * 3.1415926);
     char* ptr = ( char* )&data;
     static uint32_t loop_count = 0;
     console.printf("-------------------\r\n");
