+++ one-third-hal/core/stime-scheduler.h
@@ -72,8 +72,8 @@ typedef void (*TaskHandle)(void);
 
 typedef struct TaskCell_s {
     uint8_t run;     // 1: need to run, 0: do not need to run
-    uint16_t time;   // count down time, when 0, set run = 1
-    uint16_t ticks;  // total tick number for a loop
+    uint32_t time;   // count down time, when 0, set run = 1
+    uint32_t ticks;  // total tick number for a loop
     char name[_STIME_TASK_NAME_LEN];
     TaskHandle handle;
 } TaskCell_t;
